Adding zmq service to QT gui class
